etl/requirements.txt
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0